+++ examples/dataset/keypoint_detection/pyproject.toml
@@ -7,7 +7,8 @@ license = "Apache-2.0"
 
 [tool.poetry.dependencies]
 python = ">=3.8,<3.12"
-kolena = ">=0.100.0,<1"
+# kolena = ">=0.100.0,<1"
+kolena = { path = "../../..", develop = true }
 s3fs = "^2023.5.0"
 fsspec = "^2023.5.0"
 pandera = ">=0.9.0,<0.16"
