+++ kolena/_experimental/dataset/_evaluation.py
@@ -115,7 +115,7 @@ def fetch_results(
     """
     Fetch results given dataset name and model name.
     """
-    log.info(f"Fetching results for model '{model}' on dataset '{dataset}'")
+    log.info(f"fetching results for model '{model}' on dataset '{dataset}'")
     df = _fetch_results(dataset, model)
 
     df_datapoints = _to_deserialized_dataframe(df.drop_duplicates(subset=[COL_DATAPOINT]), column=COL_DATAPOINT)
@@ -129,7 +129,7 @@ def fetch_results(
                 _to_deserialized_dataframe(df_matched, column=COL_RESULT),
             ),
         )
-    log.info(f"Fetched results for model '{model}' on dataset '{dataset}'")
+    log.info(f"fetched results for model '{model}' on dataset '{dataset}'")
     return df_datapoints, df_results_by_eval
 
 
@@ -177,7 +177,7 @@ def test(
 
     _validate_configs([cfg for cfg, _ in results])
     for config, df_result_input in results:
-        log.info(f"start evaluation with configuration {config}" if config else "start evaluation")
+        log.info(f"uploading test results with configuration {config}" if config else "uploading test results")
         if isinstance(df_result_input, pd.DataFrame):
             validate_dataframe_ids(df_result_input, existing_dataset.id_fields)
             df_results = _process_result(config, df_result_input, existing_dataset.id_fields)
@@ -192,4 +192,4 @@ def test(
                 upload_data_frame(df=df_results, batch_size=BatchSize.UPLOAD_RECORDS.value, load_uuid=load_uuid)
 
     _upload_results(model, load_uuid, existing_dataset.id)
-    log.info(f"Uploaded test results for model '{model}' on dataset '{dataset}'")
+    log.info(f"uploaded test results for model '{model}' on dataset '{dataset}'")
