+++ kolena/_utils/endpoints.py
@@ -82,7 +82,7 @@ def _get_test_suite_url(client_state: _ClientState, test_suite_id: int) -> str:
 
 def _get_dataset_url(client_state: _ClientState, dataset_id: str) -> str:
     platform_url = _get_platform_url(client_state)
-    return f"{platform_url}/datasets?{urlencode(dict(datasetId=dataset_id))}"
+    return f"{platform_url}/dataset?{urlencode(dict(datasetId=dataset_id))}"
 
 
 def get_dataset_url(dataset_id: int) -> str:
